move transak.ts to utils
